# docs/swagger.yaml
openapi: 3.0.0
info:
  title: GearUp Tech E-Commerce API
  version: 1.0.0
  description: RESTful API สำหรับระบบ E-Commerce
servers:
  - url: https://backend-gearuptech.onrender.com/api
...
paths:
...
  /product:
    get:
      tags: [Product]
      summary: ดึงสินค้าทั้งหมด
      responses:
        '200':
          description: รายการสินค้า

    post:
      tags: [Product]
      summary: เพิ่มสินค้าใหม่
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - price
                - description
                - category
              properties:
                name:
                  type: string
                price:
                  type: number
                description:
                  type: string
                category:
                  type: string
                bestseller:
                  type: boolean
                image1:
                  type: string
                  format: binary
                image2:
                  type: string
                  format: binary
                image3:
                  type: string
                  format: binary
                image4:
                  type: string
                  format: binary
      responses:
        '201':
          description: เพิ่มสินค้าสำเร็จ
  /product/{_id}/updateproduct:
    put:
      tags: [Product]
      summary: อัปเดตข้อมูลสินค้า (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: _id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                description:
                  type: string
                category:
                  type: string
      responses:
        '200':
          description: อัปเดตสินค้าสำเร็จ
  /product/{_id}:
    delete:
      tags: [Product]
      summary: ลบสินค้า (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: _id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ลบสินค้าสำเร็จ
  /product/{_id}/bestseller:
    put:
      tags: [Product]
      summary: เพิ่มสินค้าเป็น Bestseller (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: _id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bestseller:
                  type: boolean
      responses:
        '200':
          description: อัปเดตสถานะ Bestseller สำเร็จ
    delete:
      tags: [Product]
      summary: ลบสินค้าออกจาก Bestseller (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: _id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ลบสถานะ Bestseller สำเร็จ
...
components:
  securitySchemes:
    bearerAuth:
      type: https
      scheme: bearer
      bearerFormat: JWT